工作量证明

+++ src/main/java/one/wangwei/blockchain/pow/PowResult.java
@@ -13,13 +13,13 @@ public class PowResult {
     /**
      *
      */
-    private long nonce;
+    private int nonce;
     /**
      *
      */
     private byte[] hash;
 
-    public PowResult(long nonce, byte[] hash) {
+    public PowResult(int nonce, byte[] hash) {
         this.nonce = nonce;
         this.hash = hash;
     }
